fix: resolve all static analysis warnings (CppCoreCheck)
- Remove unused variables and functions across overlay, mouse drivers, and main files
- Fix constructor field initialization order in SerialConnection and MakcuConnection
- Add parentheses for operator precedence clarity in other_tools.cpp and my_enc.cpp
- Fix size_t to int conversions with explicit casts
- Add #ifndef guard for CUDA_CHECK macro to prevent redefinition
- Add NVCC diag-suppress for __shared__ variable warning (20054)
- Fix struct initialization braces in kmboxNet.cpp

Static analysis now reports 0 warnings for project code.

+++ needaimbot/cuda/preprocessing.cu
@@ -139,7 +139,7 @@ extern "C" cudaError_t unifiedPreprocessing(
         dst_rgb_chw,
         src_bgra.cols(),
         src_bgra.rows(),
-        src_bgra.step(),
+        static_cast<int>(src_bgra.step()),
         target_width,
         target_height,
         scale_factor
